# specs/openapi/propnexus.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropNexus API",
    "version": "1.0.0",
    "description": "Generado desde el router montado (`pnpm --filter @plataforma/api docs:openapi`), no mantenido a mano — ver apps/api/scripts/generate-openapi.ts. El código de éxito se lee del `res.status(2xx)` real del handler (con el de creación si el handler tiene más de uno, caso idempotente); las respuestas no traen schema (ver el comentario del generador). Body y query sí son el schema Zod real donde existe."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local (pnpm dev)"
    }
  ],
  "paths": {
...
    "/api/v1/evidence/{id}": {
      "get": {
        "summary": "GET /api/v1/evidence/:id",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(Evidence:id → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "patch": {
        "summary": "PATCH /api/v1/evidence/:id",
        "description": "auth + autoriza(rol(admin|developer) · proyecto(Evidence:id → developer))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string",
                    "minLength": 1
                  },
                  "authoritative": {
                    "type": "boolean"
                  },
                  "evidenceType": {
                    "type": "string",
                    "enum": [
                      "document",
                      "photo",
                      "certificate"
                    ]
                  },
                  "stageId": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Body o query no pasan el schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "delete": {
        "summary": "DELETE /api/v1/evidence/:id",
        "description": "auth + autoriza(rol(admin) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Sin contenido"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/evidence/{id}/download": {
      "get": {
        "summary": "GET /api/v1/evidence/:id/download",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(Evidence:id → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/evidence/reconcile": {
      "post": {
        "summary": "POST /api/v1/evidence/reconcile",
        "description": "auth + autoriza(rol(admin) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/evidence/{id}/anchor": {
      "post": {
        "summary": "POST /api/v1/evidence/:id/anchor",
        "description": "auth + autoriza(rol(admin) · proyecto(Evidence:id → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/evidence/{bundleId}/proof/{fileHash}": {
      "get": {
        "summary": "GET /api/v1/evidence/:bundleId/proof/:fileHash",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(EvidenceBundle:bundleId → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bundleId",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "fileHash",
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/evidence/{bundleId}/files": {
      "get": {
        "summary": "GET /api/v1/evidence/:bundleId/files",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(EvidenceBundle:bundleId → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bundleId",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "formErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fieldErrors": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "formErrors",
          "fieldErrors"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}